fixes and new code for var add calculation.

+++ tools/dwarfDbg/dwarfDbgLocationInfo.h
@@ -43,7 +43,7 @@
 
 typedef uint8_t (* getAddressSizeAndMax_t)(dwarfDbgPtr_t self, Dwarf_Half *size, Dwarf_Addr *max, Dwarf_Error *err);
 typedef uint8_t (* addLocation_t)(dwarfDbgPtr_t self, char *dirName);
-typedef uint8_t (* getLocationList_t)(dwarfDbgPtr_t self, size_t dieAndChildrenIdx, size_t dieInfoIdx, Dwarf_Bool isSibling, size_t attrIdx, Dwarf_Attribute attr);
+typedef uint8_t (* getLocationList_t)(dwarfDbgPtr_t self, size_t dieAndChildrenIdx, size_t dieInfoIdx, Dwarf_Bool isSibling, int attrIdx, Dwarf_Attribute attr);
 
 typedef struct locationOp {
   Dwarf_Small op;
